fully working invite links

+++ api/submission.js
@@ -8,7 +8,7 @@ export default {
     check(req, res, next) {
       const contest = db.contests.latest()
       if (contest.status !== "submissions") return res.sendStatus(403)
-      if (db.artists.submission(contest.id, req.user.id)) return res.sendStatus(409)
+      if (db.submissions.artist(contest.id, req.user.id)) return res.sendStatus(409)
       next()
     },
     upload: {
@@ -65,8 +65,9 @@ export default {
     execute(req, res) {
       const contest = db.contests.latest()
       if (contest.status !== "submissions") return res.sendStatus(403)
-      const submission = db.artists.submission(contest.id, req.user.id)
+      const submission = db.submissions.artist(contest.id, req.user.id)
       if (!submission) return res.sendStatus(404)
+      if (submission.artists[0].id !== req.user.id) return res.sendStatus(403)
       fs.promises.unlink(`assets/images/submissions/${contest.id}/${submission.image}.png`).catch(() => {})
       fs.promises.unlink(`assets/images/submissions/${contest.id}/${submission.image}.webp`).catch(() => {})
       fs.promises.unlink(`assets/images/submissions/${contest.id}/${submission.image}_thumbnail_small.webp`).catch(() => {})
